// elza/elza-react/src/api/generated/model/ae-item-vo.ts
// tslint:disable
/**
 * CLIENT API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Hodnota atributu
 * @export
 * @interface AeItemVO
 */
export interface AeItemVO {
    /**
     * Discriminator
     * @type {string}
     * @memberof AeItemVO
     */
    '@class': string;
    /**
     *
     * @type {number}
     * @memberof AeItemVO
     */
    id?: number;
    /**
     *
     * @type {string}
     * @memberof AeItemVO
     */
    uuid?: string;
    /**
     *
     * @type {number}
     * @memberof AeItemVO
     */
    itemTypeId: number;
    /**
     *
     * @type {number}
     * @memberof AeItemVO
     */
    itemSpecId?: number;
    /**
     *
     * @type {string}
     * @memberof AeItemVO
     */
    textValue: string;
}