// elza/elza-react/src/api/generated/model/system-code.ts
// tslint:disable
/**
 * CLIENT API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Systémový typ:  * `COORDINATES` - souřadnice  * `INTEGER` - celé číslo  * `RECORD_REF` - odkaz na archivní entitu  * `BIT` - boolean  * `STRING` - jednořádkový text  * `TEXT` - text  * `NULL` - enum  * `LINK` - odkaz  * `UNITDATE` - datace
 * @export
 * @enum {string}
 */
export enum SystemCode {
    COORDINATES = 'COORDINATES',
    INTEGER = 'INTEGER',
    RECORDREF = 'RECORD_REF',
    BIT = 'BIT',
    STRING = 'STRING',
    TEXT = 'TEXT',
    NULL = 'NULL',
    LINK = 'LINK',
    UNITDATE = 'UNITDATE'
}